+++ web/modules/custom/giv_din_stemme/js/giv_din_stemme.js
@@ -17,6 +17,16 @@ await register(await connect());
   const fileElement = document.querySelector("#audio_input");
   const durationElement = document.querySelector("#recording_duration");
 
+  function hideElement(element) {
+    element.classList.add("hidden");
+    element.setAttribute("aria-hidden", true);
+  }
+
+  function showElement(element) {
+    element.classList.remove("hidden");
+    element.setAttribute("aria-hidden", false);
+  }
+
   // Initialize
   try {
     const audioStream = await navigator.mediaDevices.getUserMedia({
@@ -86,31 +96,10 @@ await register(await connect());
       });
 
       mediaRecorder.onstop = function () {
-        const clipContainer = document.createElement("article");
-        // TODO: Do we use this clipLabel?
-        const clipLabel = document.createElement("p");
-        const audio = document.createElement("audio");
-        const deleteButton = document.createElement("button");
+        const audio = soundClips.querySelector("audio");
+        const deleteButton = soundClips.querySelector("button");
 
-        soundClips.classList.remove("hidden");
-
-        clipContainer.classList.add(
-          "clip",
-          "flex",
-          "gap-2",
-          "border-t",
-          "my-1",
-          "py-2",
-        );
-        deleteButton.classList.add("btn-danger", "self-center");
-        audio.setAttribute("controls", "");
-        // TODO: This should be translateable
-        deleteButton.textContent = "Delete";
-
-        clipContainer.appendChild(audio);
-        clipContainer.appendChild(clipLabel);
-        clipContainer.appendChild(deleteButton);
-        soundClips.appendChild(clipContainer);
+        showElement(soundClips);
 
         const blob = new Blob(chunks, { type: "audio/wav" });
         chunks = [];
@@ -129,15 +118,13 @@ await register(await connect());
         // Set duration time
         durationElement.value = Math.round((endTime - startTime) / 1000);
 
-        deleteButton.onclick = (e) => {
-          let evtTgt = e.target;
-          evtTgt.parentNode.parentNode.removeChild(evtTgt.parentNode);
+        deleteButton.onclick = () => {
           fileElement.files = new DataTransfer().files;
           startTime = null;
           endTime = null;
           submitButton.disabled = finishButton.disabled = true;
           toggleButton.disabled = false;
-          soundClips.classList.add("hidden");
+          hideElement(soundClips);
         };
 
         submitButton.disabled = finishButton.disabled = false;