+++ CHANGELOG.md
@@ -8,6 +8,8 @@ See [keep a changelog] for information about writing changes to this log.
 
 ## [Unreleased]
 
+* [PR-35](https://github.com/itk-dev/giv-din-stemme/pull/35)
+  Made button translatable.
 * [PR-37](https://github.com/itk-dev/giv-din-stemme/pull/37)
   Added “Save and finish” button
 * [PR-36](https://github.com/itk-dev/giv-din-stemme/pull/36)